+++ src/main/java/com/github/zastrixarundell/torambot/Values.java
@@ -3,8 +3,6 @@ package com.github.zastrixarundell.torambot;
 import org.discordbots.api.client.DiscordBotListAPI;
 import org.joda.time.DateTime;
 
-import java.awt.image.BufferedImage;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Properties;
 
@@ -15,25 +13,23 @@ public class Values
 
     private static String version = "";
 
-    private static int userCount, guildCount, commandCount;
+    private static int userCount, guildCount;
 
     public static final String profileImageURL = "https://raw.githubusercontent.com/ZastrixArundell/ToramBot/master/images/profile.png";
-
-    public final static String footerMessage = "Support me by going on: http://corneey.com/w2ObhY";
 
     public final static String toramLogo = "https://toramonline.com/index.php?media/toram-online-logo.50/full&d=1463410056";
 
     public final static String inviteLink = "http://ceesty.com/w2Ncfe";
 
-    public final static String donationLink = "https://www.patreon.com/zastrix_arundell";
+    public final static String donationLink = "https://www.paypal.com/cgi-bin/webscr?cmd=_s-xclick&hosted_button_id=9U8Z5E9L485G2&source=url";
 
     public final static String supportLink = "https://discord.gg/MdASH22";
 
-    public final static String patreonLogo = "https://upload.wikimedia.org/wikipedia/commons/thumb/9/94/Patreon_logo.svg/541px-Patreon_logo.svg.png";
+    public final static String donationLogo = "http://pngimg.com/uploads/paypal/paypal_PNG22.png";
 
     private static DateTime lastDyeUpdate;
 
-    private static BufferedImage[] dyeImages = null;
+    private static List<String> dyeImages = null;
 
     private static String prefix = ">";
 
@@ -41,11 +37,11 @@ public class Values
 
     public static String getPrefix() { return prefix; }
 
-    public static BufferedImage[] getDyeImages() { return dyeImages; }
+    public static List<String> getDyeImages() { return dyeImages; }
 
-    public static void setDyeImages(BufferedImage[] dyeImage) { Values.dyeImages = dyeImage; }
+    public static void setDyeImages(List<String> dyeImages) { Values.dyeImages = dyeImages; }
 
-    public static void getMavenVersion()
+    static void getMavenVersion()
     {
         try
         {
@@ -66,37 +62,15 @@ public class Values
 
     public static void setLastDyeUpdate(DateTime lastDyeUpdate) { Values.lastDyeUpdate = lastDyeUpdate; }
 
-    public static int getUserCount()
-    {
-        return userCount;
-    }
+    public static int getUserCount() { return userCount; }
 
-    public static void setUserCount(int userCount)
-    {
-        Values.userCount = userCount;
-    }
+    public static void setUserCount(int userCount) { Values.userCount = userCount; }
 
-    public static int getGuildCount()
-    {
-        return guildCount;
-    }
+    public static int getGuildCount() { return guildCount; }
 
-    public static void setGuildCount(int guildCount)
-    {
-        Values.guildCount = guildCount;
-    }
+    public static void setGuildCount(int guildCount) { Values.guildCount = guildCount; }
 
-    public static int getCommandCount()
-    {
-        return commandCount;
-    }
-
-    public static void setCommandCount(int commandCount)
-    {
-        Values.commandCount = commandCount;
-    }
-
-    public static DiscordBotListAPI getApi() { return api;     }
+    public static DiscordBotListAPI getApi() { return api; }
 
     public static void setApi(DiscordBotListAPI api) { Values.api = api; }
 }