+++ src/main/java/com/github/zastrixarundell/torambot/utils/AESHelper.java
@@ -34,6 +34,4 @@ public class AESHelper
         return new String(deciphered);
     }
 
-
-
 }